+++ changelog.rst
@@ -22,6 +22,8 @@ Yet to be versioned and released. Only available from *dev* branch until then.
 
 * fix dump file update bug due to missing 'divisions' dimension
   (`#32 <https://github.com/cm4twc-org/cm4twc/issues/32>`_)
+* fix model identifier renaming not propagating to its components' identifiers
+  (`#48 <https://github.com/cm4twc-org/cm4twc/issues/48>`_)
 
 .. rubric:: Enhancements